requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.5
django-decouple==2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
multidict==6.7.0
packaging==25.0
propcache==0.4.1
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
twilio==9.8.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.22.0
